Feat#30오류처리

+++ src/main/java/com/sungkyul/decemberproject/Awesomely_Delicious/service/RestaurantService.java
@@ -4,13 +4,14 @@ import com.sungkyul.decemberproject.Awesomely_Delicious.api.RestaurantForm;
 import com.sungkyul.decemberproject.Awesomely_Delicious.api.UpdateRestaurantForm;
 import com.sungkyul.decemberproject.Awesomely_Delicious.domain.Restaurant;
 import com.sungkyul.decemberproject.Awesomely_Delicious.domain.User;
+import com.sungkyul.decemberproject.Awesomely_Delicious.exceptioncontroller.ApiException;
+import com.sungkyul.decemberproject.Awesomely_Delicious.exceptioncontroller.ExceptionEnum;
 import com.sungkyul.decemberproject.Awesomely_Delicious.repository.MemberRepository;
 import com.sungkyul.decemberproject.Awesomely_Delicious.repository.RestaurantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
 
 @Service
 @Transactional
@@ -21,7 +22,7 @@ public class RestaurantService {
 
     public void addRestaurant(RestaurantForm form){
         User user = memberRepository.findById(form.getUser_id());
-
+        if (user== null) throw new ApiException(ExceptionEnum.NonExistent_User);
         Restaurant restaurant = new Restaurant();
         restaurant.setRestaurantName(form.getRestaurant_name());
         restaurant.setMemo(form.getMemo());
@@ -33,11 +34,14 @@ public class RestaurantService {
     }
 
     public void deleteRestaurant(Long restaurantId){
+        Restaurant restaurant = restaurantRepository.findById(restaurantId);
+        if (restaurant== null) throw new ApiException(ExceptionEnum.NonExistent_Restaurant);
         restaurantRepository.deleteById(restaurantId);
     }
 
     public void updateRestaurant(UpdateRestaurantForm form, Long restaurantId){
         Restaurant restaurant = restaurantRepository.findById(restaurantId);
+        if (restaurant== null) throw new ApiException(ExceptionEnum.NonExistent_Restaurant);
         restaurant.setStarCount(form.getStar_count());
         restaurant.setMemo(form.getMemo());
     }
